requirements.txt
honeybee-core>=1.10.0
honeybee-radiance-command>=1.0.0
